# openapi.yaml
openapi: 3.0.3
info:
  title: CEC HTTP Bridge API
  description: |
    REST API for controlling HDMI-CEC devices (Rust implementation). Power,
    volume, source/HDMI switching, remote keys, raw CEC commands, MQTT
    settings, a dev console surface, and real-time events (SSE + WebSocket).
    Intended for home automation, media centers, and scripting.

    ## Envelope
    Every JSON response uses the envelope
    `{"status":"success"|"error","message":"...","data":...}`.
    `message` and `data` are omitted when empty/absent. Long-running
    operations return `success` with `data.accepted=true` (HTTP 202).

    ## Authentication
    Auth is **optional** and active only when the server was started with a
    token (`-token <secret>` / `auth_token` in `config.json`). When active,
    every `/api` route **except `GET /api/health`** (and `GET /metrics`)
    requires the token via `Authorization: Bearer <token>`,
    `X-Auth-Token: <token>`, `?key=<token>`, or the `capi_token` cookie set by
    the UI login page. Mutating cookie-authenticated requests additionally
    require same-origin. When no token is configured, no auth is required.
  version: 2.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local (default)
  - url: http://{host}:8080
    description: Network
    variables:
      host:
        default: localhost
security:
  - BearerAuth: []
...
paths:
...
  /api/volume/up:
    post:
      tags: [Volume]
      summary: Volume up
      description: >
        Volume up via libcec (targets the TV's system audio path). Use the
        addressed variant to target another device (e.g. 5 = audio system).
      operationId: volumeUp
      responses:
        '200':
          description: Volume up sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/volume/down:
    post:
      tags: [Volume]
      summary: Volume down
      operationId: volumeDown
      responses:
        '200':
          description: Volume down sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        Optional. Active only when the server runs with a token
        (`-token <secret>`). `X-Auth-Token`, `?key=`, and the `capi_token`
        cookie are accepted equivalently. When no token is configured this
        security requirement is a no-op.
...
  schemas:
    ApiResponse:
      type: object
      description: >
        Standard envelope. `message` omitted when empty; `data` omitted when
        absent. Async acceptance returns success with data.accepted=true.
      required: [status]
      properties:
        status:
          type: string
          enum: [success, error]
        message:
          type: string
          description: Human-readable message (omitted when empty)
        data:
          description: Response payload (varies by endpoint; omitted when absent)
          nullable: true
...
  responses:
    BadRequest:
      description: Invalid parameters or request body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            status: error
            message: Invalid logical address
    InternalError:
      description: CEC operation failed or server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            status: error
            message: failed to open adapter